Create idp package, sketch out new entrypoint
To do multi-tenant, we want to be able to run multiple instances of the IDP.
This starts that, along with re-structuring the main a bit. The e2e tests are
now a bit more black-box too.
