+++ e2e/e2e_test.go
@@ -1,4 +1,4 @@
-package main
+package e2e_test
 
 import (
 	"context"
@@ -8,6 +8,7 @@ import (
 	"net"
 	"net/url"
 	"os"
+	"path/filepath"
 	goruntime "runtime"
 	"strconv"
 	"testing"
@@ -22,6 +23,7 @@ import (
 	"github.com/lstoll/oidc/clitoken"
 	"github.com/lstoll/oidc/core/staticclients"
 	dbpkg "github.com/lstoll/webauthn-oidc-idp/db"
+	"github.com/lstoll/webauthn-oidc-idp/internal/idp"
 	"github.com/lstoll/webauthn-oidc-idp/internal/queries"
 	_ "github.com/mattn/go-sqlite3"
 	"golang.org/x/oauth2"
@@ -93,15 +95,12 @@ func TestE2E(t *testing.T) {
 		t.Fatal(err)
 	}
 
-	issConfig := issuerConfig{
-		URL: issU,
-		Clients: []staticclients.Client{
-			{
-				ID:                      "test-cli",
-				Secrets:                 []string{"public"},
-				Public:                  true,
-				PermitLocalhostRedirect: true,
-			},
+	clients := []staticclients.Client{
+		{
+			ID:                      "test-cli",
+			Secrets:                 []string{"public"},
+			Public:                  true,
+			PermitLocalhostRedirect: true,
 		},
 	}
 
@@ -120,7 +119,7 @@ func TestE2E(t *testing.T) {
 
 	serveErr := make(chan error, 1)
 	go func() {
-		if err := serve(serveCtx, sqldb, db, issConfig, net.JoinHostPort("localhost", port), ""); err != nil {
+		if err := idp.ServeCmd(serveCtx, sqldb, db, issU, clients, net.JoinHostPort("localhost", port), ""); err != nil {
 			serveErr <- err
 		}
 	}()
@@ -134,7 +133,7 @@ func TestE2E(t *testing.T) {
 		t.Fatal("server startup timed out")
 	}
 
-	provider, err := oidc.DiscoverProvider(ctx, issConfig.URL.String(), nil)
+	provider, err := oidc.DiscoverProvider(ctx, issU.String(), nil)
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -144,9 +143,6 @@ func TestE2E(t *testing.T) {
 		Endpoint:     provider.Endpoint(),
 		Scopes:       []string{oidc.ScopeOpenID},
 	}
-
-	// not strictly needed for the E2E tests.
-	reloadDB(net.JoinHostPort("localhost", port))
 
 	/* enable the virtual webauthn environment */
 	var virtAuthenticatorID cdpwebauthn.AuthenticatorID
@@ -194,9 +190,9 @@ func TestE2E(t *testing.T) {
 			WebauthnHandle: must(uuid.NewRandom()),
 		}
 		if err := queries.New(sqldb).CreateUser(ctx, params); err != nil {
-			fatalf("create user: %v", err)
+			t.Fatalf("create user: %v", err)
 		}
-		ep := registrationURL(issConfig.URL, params.ID.String(), params.EnrollmentKey.String)
+		ep := idp.RegistrationURL(issU, params.ID.String(), params.EnrollmentKey.String)
 
 		runErrC := make(chan error, 1)
 		doneC := make(chan struct{}, 1)
@@ -452,4 +448,21 @@ func (c *chromeDPOpener) Open(ctx context.Context, url string) error {
 		c.notifyCh <- struct{}{}
 	}
 	return nil
+}
+
+func openTestDB(t *testing.T) *idp.DB {
+	t.Helper()
+
+	db, err := idp.OpenDB(filepath.Join(t.TempDir(), "db.json"))
+	if err != nil {
+		t.Fatalf("open database: %v", err)
+	}
+	return db
+}
+
+func must[T any](v T, err error) T {
+	if err != nil {
+		panic(err)
+	}
+	return v
 }
