+++ internal/idp/sessions.go
@@ -1,4 +1,4 @@
-package main
+package idp
 
 import (
 	"time"
